% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/octad.db.R
\docType{package}
\name{octad.db}
\alias{octad.db}
\title{Open Cancer TherApeutic Discovery (OCTAD) database package}
\description{
This is a support package for the main package \code{octad} which can be obtained \href{https://bioconductor.org/packages/octad}{there}
}
\details{
Package includes all required data for drug repurposing OCTAD pipeline.
Initialization of the pipeline starts with listing all available samples:

\code{phenoDF=.eh[["EH7274"]]} #load data.frame with samples included in the OCTAD database.
\code{head(phenoDF)} #list all data included within the package\preformatted{Besides, the package includes examples of the output from diffExp and runsRGES functions from the octad package along with desctiption of the data:
```sRGES=octad.db::res_example``` #load example res from octad.db
```?res_example```
```sRGES=octad.db::sRGES_example``` #load example sRGES from octad.db
```?sRGES_example```
}

The code can be viewed at the GitHub repository,
which also lists the contributor code of conduct:

\url{https://bioconductor.org/packages/octad} or \url{https://github.com/Bin-Chen-Lab/OCTAD} for the pipeline package
}
\references{
Zeng, B., Glicksberg, B.S., Newbury, P., Chekalin, E., Xing, J., Liu, K., Wen, A., Chow, C. and Chen, B., 2021. OCTAD: an open workspace for virtually screening therapeutics targeting precise cancer patient groups using gene expression features. Nature protocols, 16(2), pp.728-753.
\url{https://www.nature.com/articles/s41596-020-00430-z}
}
